+++ src/ATK/structures/Image.py
@@ -6,6 +6,7 @@ from astropy.io.fits import ImageHDU
 from astropy.time import Time
 from astropy.units import Quantity
 from astropy.wcs import WCS
+
 from .structures_core import Container