+++ src/ATK/structures/methods/spectrum/fitting.py
@@ -1,19 +1,22 @@
-import warnings
+from __future__ import annotations
 
-import astropy.units as u
+import copy
+import warnings
+from typing import TYPE_CHECKING
+
 import numpy as np
 import pandas as pd
 from astropy.stats import sigma_clip
-from astropy.units import Quantity
-from bokeh.io import show
 from bokeh.models import ColumnDataSource, HoverTool
 from bokeh.plotting import figure
-from scipy.ndimage import gaussian_filter1d
+from scipy.ndimage import gaussian_filter1d, generic_filter, median_filter
 from scipy.optimize import OptimizeWarning, curve_fit
 from scipy.signal import find_peaks
 
 from ....plotting.formatting import format_plot
-from ....structures.Spectrum import Spectrum
+
+if TYPE_CHECKING:
+    from ....structures.Spectrum import Spectrum
 
 warnings.simplefilter("ignore", category=OptimizeWarning)
 
@@ -22,15 +25,13 @@ DEBUG = False
 C_KMS = 299792.458
 GAUSS_PARAMS = ("h", "a", "mu", "sigma")
 
-INIT_WINDOW_WIDTH = 75  # initial window width in pixels
-MIN_PROM = 1.0  # minimum prominence of peaks
-MIN_SEP = 10  # minimum separation of 'new' peaks in pixels
-MIN_WIDTH = 10
+MIN_SEP = 10  # minimum separation below which peaks are discarded
+MIN_WIDTH = 10  # minimum width of a peak in pixels
 
 PROM_WEIGHT = 2.0  # weighting on prominence in scoring peaks
 SEP_WEIGHT = 1.0  # weighting on separation in scoring peaks
 
-INIT_LOCAL_WIN_WIDTH = 200  # width of initial local window in pixels
+INIT_LOCAL_WIN_WIDTH = 200  # width of initial local window in wavelength
 FINAL_LOCAL_WIN_WIDTH = 5  # width of final local window in sigma
 
 CORE_SIGMA_WIDTH = 3  # width of 'core' of peak in sigma, used to determine edge pixels
@@ -39,15 +40,15 @@ MIN_EDGE_POINTS = 4  # minimum number of points to use in continuum estimation
 FINAL_PLOTTING_WIDTH = 4  # width of final plotted Gaussian in sigma
 
 
-def get_velocities(wav: Quantity, wav_ref: Quantity) -> np.ndarray:
-    return (wav - wav_ref) / wav_ref * C_KMS * u.Unit("km s-1")
-
-
-def gaussian(x, h, a, mu, sigma, sign=-1.0):
+def gaussian(x: np.ndarray, h: float, a: float, mu: float, sigma: float, sign: int = -1):
     return h + sign * a * np.exp(-0.5 * ((x - mu) / sigma) ** 2)
 
 
 def make_gaussian(fixed):
+    """
+    Models a Gaussian with or without fixed parameters (h/a/mu/sigma)
+    """
+
     def model(x, *free):
         params = {}
         i = 0
@@ -62,16 +63,20 @@ def make_gaussian(fixed):
     return model
 
 
-def estimate_continuum(plot, x, y, mu, sigma):
+def estimate_continuum(plot: figure, x: np.ndarray, y: np.ndarray, mu: float, sigma: float):
+    """
+    Estimates the local continuum by masking out a feature
+    """
+
     wing_mask = np.abs(x - mu) > CORE_SIGMA_WIDTH * sigma
 
     x_w = x[wing_mask]
     y_w = y[wing_mask]
 
-    clipped = sigma_clip(y_w, sigma=3.0)
-
     if wing_mask.sum() < MIN_EDGE_POINTS:
-        return np.median(y), x_w[~clipped.mask], y_w[~clipped.mask]
+        return np.median(y), x_w, y_w
+
+    clipped = sigma_clip(y_w, sigma=3.0)
 
     if clipped.count() < MIN_EDGE_POINTS:
         return np.median(y_w), x_w[~clipped.mask], y_w[~clipped.mask]
@@ -79,78 +84,81 @@ def estimate_continuum(plot, x, y, mu, sigma):
     return np.median(clipped.data[~clipped.mask]), x_w[~clipped.mask], y_w[~clipped.mask]
 
 
-def fit_gaussian(plot: figure, spectrum: Spectrum, p0: dict, snr_limit: float, fixed: dict | None = None):
-    fixed = fixed or {}
+def fit_gaussian(plot: figure, spectrum: Spectrum, p0: dict, snr_limit: float):
+    """
+    Fits Gaussians to spectral features in two stages
+    """
 
+    # get width of one sample (pixel)
     d_pix = np.mean(np.diff(spectrum.wavelength))
 
+    # define initial locoal window
     init_pad = INIT_LOCAL_WIN_WIDTH * d_pix
     local_mask = (spectrum.wavelength >= p0["mu"] - init_pad) & (spectrum.wavelength <= p0["mu"] + init_pad)
     local_x, local_y = spectrum.wavelength[local_mask], spectrum.flux[local_mask]
 
-    # if DEBUG:
-    #     box = BoxAnnotation(left=p0["mu"] - INIT_PAD, right=p0["mu"] + INIT_PAD, fill_alpha=0.15, fill_color="orange")
-    #     plot.add_layout(box)
-
+    # set up priors
     p0["sigma"] = 3 * d_pix
     p0["h"] = np.median(local_y)
+    p0["a"] = np.abs(p0["h"] - local_y.min()) if p0["sign"] == -1 else np.abs(local_y.max() - p0["h"])
 
-    if p0["sign"] == -1:
-        p0["a"] = np.abs(p0["h"] - local_y.min())
-    else:
-        p0["a"] = np.abs(local_y.max() - p0["h"])
+    # set up basic Gaussian model
+    model = make_gaussian({})
+    free_p0 = [p0[k] for k in GAUSS_PARAMS]
 
-    model = make_gaussian(fixed)
-    free_p0 = [p0[k] for k in GAUSS_PARAMS if k not in fixed]
-
+    # get initial fit parameters
     popt, pcov = curve_fit(model, local_x, local_y, p0=free_p0, maxfev=20000)
-
     h, a, mu, sigma = popt
 
+    # get peak width
     sigma = abs(sigma)
     if not np.isfinite(sigma) or sigma < d_pix:
         sigma = FINAL_LOCAL_WIN_WIDTH * d_pix
 
+    # set up secondary window as multiple of peak width
     final_pad = FINAL_LOCAL_WIN_WIDTH * sigma
     local_mask = (spectrum.wavelength >= mu - final_pad) & (spectrum.wavelength <= mu + final_pad)
     local_x, local_y = spectrum.wavelength[local_mask], spectrum.flux[local_mask]
 
-    # if DEBUG:
-    #     box = BoxAnnotation(left=mu - FINAL_PAD, right=mu + FINAL_PAD, fill_alpha=0.15, fill_color="red")
-    #     plot.add_layout(box)
-
+    # create model from fit parameters and subtract from local flux
     model_free = make_gaussian({})
     y_model = model_free(local_x, h, a, mu, sigma)
     y_resid = local_y - (y_model - h)
 
+    # estimate continuum and get a new model with fixed height
     continuum_h, wing_x, wing_y = estimate_continuum(plot, local_x, y_resid, mu, sigma)
     model = make_gaussian(fixed={"h": continuum_h})
 
+    # set up other priors
     a0 = a
     mu0 = mu
     sigma0 = max(sigma, d_pix)
 
+    # get final fit
     popt, pcov = curve_fit(model, local_x, local_y, p0=[a0, mu0, sigma0], maxfev=20000)
 
+    # set up params dict
     params = {"h": continuum_h}
-
     for name, val in zip(("a", "mu", "sigma"), popt):
         params[name] = val
 
+    # clip final x for plotting and get final y
     x_mask = (spectrum.wavelength >= mu - FINAL_PLOTTING_WIDTH * sigma) & (spectrum.wavelength <= mu + FINAL_PLOTTING_WIDTH * sigma)
     final_x = spectrum.wavelength[x_mask]
     final_y = model(final_x, *popt)
 
+    # smooth local y and subtract, use MAD to estimate local noise
     local_y_smooth = gaussian_filter1d(local_y, sigma=5)
     resid = local_y - local_y_smooth
     noise = 1.4826 * np.median(np.abs(resid - np.median(resid)))
 
+    # get filtering parameters
     peak_amp = final_y.max() - final_y.min()
     peak_width = final_x.max() - final_x.min()
     peak_pix = len(final_x)
-
     peak_snr = peak_amp / noise
 
+    # used for HoverTool when DEBUG is True
     filtering = {
         "peak_width": [peak_width],
         "peak_amp": peak_amp,
@@ -161,6 +169,7 @@ def fit_gaussian(plot: figure, spectrum: Spectrum, p0: dict, snr_limit: float, f
         "continuum_h": continuum_h,
     }
 
+    # discard peaks that cover far too much of the spectrum
     if peak_width > 0.1 * (spectrum.wavelength.max() - spectrum.wavelength.min()):
         return None, f"width too wide ({peak_width:.2f})", filtering
 
@@ -168,9 +177,11 @@ def fit_gaussian(plot: figure, spectrum: Spectrum, p0: dict, snr_limit: float, f
         plot.line(local_x, [continuum_h] * len(local_x), line_color="red", legend_label="Continuum Level")
         plot.scatter(wing_x, wing_y, size=4, alpha=0.6, color="orange", legend_label="Continuum Points")
 
+    # discard really thin peaks
     if peak_width < 10:
         return None, f"width too narrow ({peak_pix} px)", filtering
 
+    # discard peaks with poor snr
     if peak_snr < snr_limit:
         return None, f"snr too low ({peak_snr:.2f})", filtering
 
@@ -188,6 +199,10 @@ def fit_gaussian(plot: figure, spectrum: Spectrum, p0: dict, snr_limit: float, f
 
 
 def detect_features(plot, spectrum: Spectrum, min_prominence: float, smoothing: int):
+    """
+    Identifies features in a spectrum
+    """
+
     flux_smooth = gaussian_filter1d(spectrum.flux, sigma=smoothing)
 
     # emission peaks
@@ -202,6 +217,7 @@ def detect_features(plot, spectrum: Spectrum, min_prominence: float, smoothing: 
     prominences = np.concatenate([em_props["prominences"], abs_props["prominences"]])
     widths = np.concatenate([em_props["widths"], abs_props["widths"]])
 
+    # generate score from prominence and width
     score = PROM_WEIGHT * prominences + SEP_WEIGHT * widths
 
     data = pd.DataFrame({"idx": peaks, "sign": signs, "prominence": prominences, "width": widths, "score": score}).sort_values("idx")
@@ -214,6 +230,7 @@ def detect_features(plot, spectrum: Spectrum, min_prominence: float, smoothing: 
             last_idx = row["idx"]
             continue
 
+        # keep only best of two close peaks
         if row["idx"] - last_idx < MIN_SEP:
             if row["score"] > keep[-1]["score"]:
                 keep[-1] = row
@@ -226,10 +243,14 @@ def detect_features(plot, spectrum: Spectrum, min_prominence: float, smoothing: 
     data["vel"] = spectrum.wavelength[data["idx"].to_numpy(dtype=int)]
     data["flux"] = spectrum.flux[data["idx"].to_numpy(dtype=int)]
 
-    return flux_smooth, data
+    return data
 
 
 def subtract_peak(wavelength: np.ndarray, flux: np.ndarray, model_x: np.ndarray, model_y: np.ndarray, h: float):
+    """
+    Subtracts a peak from a spectrum given its height
+    """
+
     mask = (wavelength >= model_x.min()) & (wavelength <= model_x.max())
     flux_new = flux.copy()
     flux_new[mask] -= model_y - h
@@ -242,26 +263,35 @@ def mad_sigma(x):
 
 
 def remove_false_features(plot, spectrum, fits):
-    from scipy.ndimage import generic_filter, median_filter
+    """
+    Removes false features created by the intersection of two close real features
+    """
 
+    # subtract features from spectrum
     resid_flux = spectrum.flux.copy()
     for fit in fits:
         resid_flux = subtract_peak(spectrum.wavelength, resid_flux, fit["model_x"], fit["model_y"], fit["h"])
 
+    # sigma clip
     resid_flux = sigma_clip(resid_flux, sigma=3)
     resid_flux = resid_flux.filled(np.nan)
     resid_flux = np.asarray(resid_flux, dtype=float)
 
+    # running median window
     window = 200
     continuum = median_filter(resid_flux, size=window)
 
+    # running Median Absolute Deviation window to get width of continuum
     sigma = generic_filter(resid_flux, mad_sigma, size=window)
 
     keep = []
     for fit in fits:
         idx = np.abs(spectrum.wavelength - fit["mu"]).argmin()
+
+        # get value at peak of feature
         peak_val = fit["model_y"].max() if fit["sign"] == 1 else fit["model_y"].min()
 
+        # filter out features where the peak value is close to the continuum
         if np.abs(peak_val - continuum[idx]) > 3 * sigma[idx]:
             keep.append(fit)
 
@@ -272,42 +302,54 @@ def remove_false_features(plot, spectrum, fits):
     return keep
 
 
-def do_fitting(spectrum: Spectrum, prominence: float, snr: float, smoothing: int, debug=False):
+def do_fitting(plot: figure, spectrum: Spectrum, prominence: float = 2, smoothing: int = 3, snr: float = 3, **kwargs):
+    """
+    Identifies and fits any number of spectral absorption and emission features
+    """
+
+    # set debugging mode
     global DEBUG
-    DEBUG = debug
+    DEBUG = kwargs.get("debug", False)
 
-    wavelength_unit = spectrum.wavelength.unit
+    # convert quantity arrays to basic arrays
+    spectrum = copy.deepcopy(spectrum)
+    spectrum.wavelength = spectrum.wavelength.copy().value
+    spectrum.flux = spectrum.flux.copy().value
 
-    spectrum.wavelength = spectrum.wavelength.value
-    spectrum.flux = spectrum.flux.value
+    # get features
+    peak_data = detect_features(plot, spectrum, prominence, smoothing)
 
-    plot = figure(width=1000, height=500, x_axis_label=f"Wavelength / {wavelength_unit.to_string('unicode')}", y_axis_label="Flux")
-    plot.line(spectrum.wavelength, spectrum.flux, line_color="black", line_alpha=0.5)
+    flux_smooth = gaussian_filter1d(spectrum.flux, sigma=smoothing)
 
-    flux_smooth, peak_data = detect_features(plot, spectrum, prominence, smoothing)
-
-    fits, responses = [], []
-    filtering = []
+    # loop through returned peaks
+    fits, responses, filtering = [], [], []
     for index, row in peak_data.iterrows():
+        # assemble priors
         p0 = {}
         p0["mu"] = row["vel"]
         p0["sign"] = row["sign"]
 
-        fit, response, filter = fit_gaussian(plot, spectrum, p0, snr_limit=snr, fixed={})
+        # fit gaussians to features
+        fit, response, filter = fit_gaussian(plot, spectrum, p0, snr_limit=snr)
 
         if fit:
             fits.append(fit)
+
         responses.append(response)
         filtering.append(filter)
 
+    # remove fake features that result from two close real features
     fits = remove_false_features(plot, spectrum, fits)
+
+    # plot peaks
     for fit in fits:
         plot.line(fit["model_x"], fit["model_y"], line_color="limegreen", line_width=2, legend_label="Detected Peaks")
 
+    # plot smoothed flux
     plot.line(spectrum.wavelength, flux_smooth, line_alpha=0.3, line_width=3, line_color="red", legend_label="Smoothed Flux")
 
+    # peak filtering overlay
     filtering_df = pd.concat(pd.DataFrame(dct) for dct in filtering).reset_index(drop=True)
-
     if DEBUG:
         df = pd.DataFrame(
             {
@@ -325,4 +367,4 @@ def do_fitting(spectrum: Spectrum, prominence: float, snr: float, smoothing: int
 
     plot = format_plot("spectrum", plot)
 
-    show(plot)
+    return plot